Move parse_version into ddtrace.utils.version
