+++ ddtrace/contrib/flask/patch.py
@@ -10,6 +10,7 @@ from ...constants import ANALYTICS_SAMPLE_RATE_KEY
 from ...constants import SPAN_MEASURED_KEY
 from ...ext import SpanTypes
 from ...internal.logger import get_logger
+from ...utils.version import parse_version
 from ...utils.wrappers import unwrap as _u
 from .helpers import get_current_app
 from .helpers import simple_tracer
@@ -53,7 +54,7 @@ config._add(
 #      (0, 9, 0) != (0, 9)
 #      (0, 8, 5) <= (0, 9)
 flask_version_str = getattr(flask, "__version__", "0.0.0")
-flask_version = trace_utils.parse_version(flask_version_str)
+flask_version = parse_version(flask_version_str)
 
 
 def patch():
