+++ ddtrace/contrib/molten/patch.py
@@ -13,6 +13,7 @@ from ...internal.compat import urlencode
 from ...utils.formats import asbool
 from ...utils.formats import get_env
 from ...utils.importlib import func_name
+from ...utils.version import parse_version
 from ...utils.wrappers import unwrap as _u
 from .wrappers import MOLTEN_ROUTE
 from .wrappers import WrapperComponent
@@ -21,7 +22,7 @@ from .wrappers import WrapperRenderer
 from .wrappers import WrapperRouter
 
 
-MOLTEN_VERSION = trace_utils.parse_version(molten.__version__)
+MOLTEN_VERSION = parse_version(molten.__version__)
 
 # Configure default configuration
 config._add(
