+++ ddtrace/contrib/trace_utils.py
@@ -305,18 +305,3 @@ def flatten_dict(
         else:
             flat[p] = v
     return flat
-
-
-def parse_version(version):
-    # type: (str) -> Tuple[int, int, int]
-    """Convert a version string to a tuple of (major, minor, micro)"""
-    # If we have any spaces/extra text, grab the first part
-    #   "1.0.0 beta1" -> "1.0.0"
-    #   "1.0.0" -> "1.0.0"
-    if " " in version:
-        version = version.split()[0]
-
-    # DEV: Do not use `packaging.version.parse`, we do not want a LegacyVersion here
-    # This will raise an InvalidVersion exception if it cannot be parsed
-    _version = packaging.version.Version(version)
-    return (_version.major, _version.minor, _version.micro)
