+++ riotfile.py
@@ -152,6 +152,7 @@ venv = Venv(
                     pkgs={
                         "msgpack": latest,
                         "attrs": ["==19.2.0", latest],
+                        "packaging": ["==17.1", latest],
                     },
                 )
             ],
