+++ setup.py
@@ -2,13 +2,14 @@ import os
 import platform
 import sys
 
-from setuptools import setup, find_packages, Extension
-from setuptools.command.test import test as TestCommand
-
 # ORDER MATTERS
 # Import this after setuptools or it will fail
 from Cython.Build import cythonize  # noqa: I100
 import Cython.Distutils
+from setuptools import Extension
+from setuptools import find_packages
+from setuptools import setup
+from setuptools.command.test import test as TestCommand
 
 
 HERE = os.path.dirname(os.path.abspath(__file__))
@@ -56,8 +57,9 @@ class Tox(TestCommand):
 
     def run_tests(self):
         # import here, cause outside the eggs aren't loaded
-        import tox
         import shlex
+
+        import tox
 
         args = self.tox_args
         if args:
@@ -166,7 +168,7 @@ setup(
             "enum34; python_version<'3.4'",
             "funcsigs>=1.0.0; python_version=='2.7'",
             "typing; python_version<'3.5'",
-            "packaging>=20.0",
+            "packaging>=17.1",
             "protobuf>=3",
             "tenacity>=5",
             "attrs>=19.2.0",
