+++ tests/tracer/test_utils.py
@@ -4,6 +4,7 @@ import unittest
 import warnings
 
 import mock
+import packaging
 import pytest
 
 from ddtrace.utils import ArgumentError
@@ -18,6 +19,7 @@ from ddtrace.utils.formats import asbool
 from ddtrace.utils.formats import get_env
 from ddtrace.utils.formats import parse_tags_str
 from ddtrace.utils.importlib import func_name
+from ddtrace.utils.version import parse_version
 
 
 class TestUtils(unittest.TestCase):
@@ -407,3 +409,31 @@ def test_cachedmethod():
             return expensive(key)
 
     cached_test_recipe(expensive, Foo().cheap, witness, cache_size)
+
+
+@pytest.mark.parametrize(
+    "version_str,expected",
+    [
+        ("5", (5, 0, 0)),
+        ("0.5", (0, 5, 0)),
+        ("0.5.0", (0, 5, 0)),
+        ("1.0.0", (1, 0, 0)),
+        ("1.2.0", (1, 2, 0)),
+        ("1.2.8", (1, 2, 8)),
+        ("2.0.0rc1", (2, 0, 0)),
+        ("2.0.0-rc1", (2, 0, 0)),
+        ("2.0.0 here be dragons", (2, 0, 0)),
+    ],
+)
+def test_parse_version(version_str, expected):
+    # type: (str, typing.Tuple[int, int, int]) -> None
+    """Ensure parse_version helper properly parses versions"""
+    assert parse_version(version_str) == expected
+
+
+def test_parse_version_errors():
+    # type: () -> None
+    """Ensure parse_version helper raises on bad input"""
+    for version_str in ["beta 1.0.0", "no version found", ""]:
+        with pytest.raises(packaging.version.InvalidVersion):
+            parse_version(version_str)
